Changes App, components and Navigation
