+++ AgroSense/App.js
@@ -1,16 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import { View, ActivityIndicator } from 'react-native';
-import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from '@react-navigation/stack';
 import { onAuthStateChanged } from 'firebase/auth';
-import { auth } from './src/app/services/firebase';  // Asegúrate de que la ruta sea correcta
-
-import WelcomeScreen from './src/app/screens/welcomeScreen'; // Corrige la ruta si es necesario
-import RegisterScreen from './src/app/screens/registerScreen'; // Corrige la ruta si es necesario
-import LoginScreen from './src/app/screens/loginScreen'; // Corrige la ruta si es necesario
-import Home from './src/app/screens/Home'; // Corrige la ruta si es necesario
-
-const Stack = createStackNavigator();
+import { auth } from './src/app/services/firebase';  // Ruta correcta a firebase.js
+import AppNavigator from './src/app/navigation/appNavigator';  // Importar la navegación desde AppNavigator.js
 
 export default function App() {
   const [loading, setLoading] = useState(true); // Estado de carga
@@ -37,20 +29,6 @@ export default function App() {
   }
 
   return (
-    <NavigationContainer>
-      <Stack.Navigator>
-        {user ? (
-          // Si el usuario está autenticado, redirige a la pantalla Home
-          <Stack.Screen name="Home" component={Home} options={{ headerShown: false }} />
-        ) : (
-          <>
-            {/* Si no está autenticado, mostrar Welcome, Register y Login */}
-            <Stack.Screen name="Welcome" component={WelcomeScreen} options={{ headerShown: false }} />
-            <Stack.Screen name="Register" component={RegisterScreen} options={{ headerShown: false }} />
-            <Stack.Screen name="Login" component={LoginScreen} options={{ headerShown: false }} />
-          </>
-        )}
-      </Stack.Navigator>
-    </NavigationContainer>
+    <AppNavigator user={user} />  // Pasar el estado del usuario a la navegación
   );
 }
