+++ AgroSense/src/app/navigation/appNavigator.js
@@ -1,0 +1,33 @@
+import React from 'react';
+import { createStackNavigator } from '@react-navigation/stack';
+import { NavigationContainer } from '@react-navigation/native';
+
+// Importar las pantallas
+import WelcomeScreen from '../screens/welcomeScreen';
+import RegisterScreen from '../screens/registerScreen';
+import LoginScreen from '../screens/loginScreen';
+import HomeScreen from '../screens/Home'  // Asegurarse de tener una pantalla HomeScreen correctamente configurada
+
+const Stack = createStackNavigator();
+
+const AppNavigator = ({ user }) => {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator>
+        {user ? (
+          // Si el usuario está autenticado, redirige a la pantalla Home
+          <Stack.Screen name="Home" component={HomeScreen} options={{ headerShown: false }} />
+        ) : (
+          <>
+            {/* Si no está autenticado, mostrar Welcome, Register y Login */}
+            <Stack.Screen name="Welcome" component={WelcomeScreen} options={{ headerShown: false }} />
+            <Stack.Screen name="Register" component={RegisterScreen} options={{ headerShown: false }} />
+            <Stack.Screen name="Login" component={LoginScreen} options={{ headerShown: false }} />
+          </>
+        )}
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+};
+
+export default AppNavigator;
